HRRR - FileLoader - Remove re-indexing the X and Y values.
Remove re-indexing the Xarray dimensions of X and Y data object that
multiplied those values by the pixel width. This information is not used
other than renaming the columns of the converted Pandas dataframe, which
can handle the formatting there based of the string instead of dropping
and re-adding an index on the Xarray object.